requirements.txt
Django==1.11.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
Django==1.11.4
django-ckeditor==5.3.0
django-js-asset==0.1.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pytz==2017.2
